Refactor Реализована заглушка на Эндпоинт на ("/api/post")
