--- src/main/java/main/api/post/ApiPostController.java
@@ -1,9 +1,0 @@
-package main.api.post;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-@RequestMapping("/api/post/")
-public class ApiPostController {
-}
