+++ src/main/java/main/controllers/GlobalController.java
@@ -1,5 +1,7 @@
-package main.api.global;
+package main.controllers;
 
+import main.api.global.BlogDataResponse;
+import main.api.global.GlobalService;
 import main.dao.entities.GlobalSettings;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.web.bind.annotation.GetMapping;
